Renaming based on PR comments
