+++ src/components/lineChart/installationselector.module.scss
@@ -4,7 +4,7 @@
   justify-content: flex-end;
 }
 
-.rwziButton {
+.installationButton {
   min-width: 15em;
   border: 1px solid lightgrey;
   padding: 0.5em;
@@ -16,7 +16,7 @@
   background-position: right 0.2em top 50%, 0 0;
 }
 
-.rwziSelector {
+.installationSelector {
   min-width: 15em;
   border-color: lightgrey;
   appearance: none;
