+++ src/components/lineChart/lineChart.module.scss
@@ -35,13 +35,13 @@ $breakpoint: 1200px;
   }
 }
 
-.legenda li {
+.legend li {
   list-style: none;
   display: flex;
   flex-direction: row;
 }
 
-.legendaMarker {
+.legendMarker {
   height: 0.5em;
   width: 0.5em;
   margin-right: 0.5em;
